Passthrough nodes with iomems only
Devices with iomems only also deserve to be passed through.
But some nodes with 'reg' properties may not describe devices.
In order to filter out major number of non-device nodes, set a
rule to ignore nodes with 'reg' properties size less than two
words. Because two words is a minimal possible description of
the iomem range.

Signed-off-by: Andrii Anisov <andrii_anisov@epam.com>